Move Cr2Header to a new .cpp file
